+++ src/components/table/index.ts
@@ -15,4 +15,5 @@ export type {
   TableBulkAction,
   TableBulkActionsClassNames,
   TableBulkActionsProps,
+  TableRowAction,
 } from "../../types/table";